Versao Vaadin 14 Java 8

+++ src/main/java/com/caracore/crm/data/repository/StatusRepository.java
@@ -1,8 +1,9 @@
 package com.caracore.crm.data.repository;
 
-import com.caracore.crm.data.entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Long> {
+import com.caracore.crm.data.entity.Status;
+
+public interface StatusRepository extends JpaRepository<Status, Integer> {
 
 }
